requirements.txt
# Python dependencies for Briefing + Status.json Integration
# Sprint 3 - Hardening (pinned versions)

# Core dependencies (MkDocs)
mkdocs==1.6.1
mkdocs-material==9.6.21
pyyaml==6.0.3

# Integration dependencies (JSON rendering + validation)
jinja2==3.1.4
jsonschema==4.23.0

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0

# Note: Chart.js se carga vía CDN (no requiere instalación Python)
# CDN: https://cdn.jsdelivr.net/npm/chart.js@4.4.0/dist/chart.umd.min.js
